refactor: remove yaml dependency and enhance documentation
Removed the 'yaml' package from the list of imports in DESCRIPTION file as it's no longer needed. Also, enhanced the documentation in gmhleasr-package.Rd by adding useful links for users. In utils_schemas.R, removed a line that was reading a yaml file which is not required anymore and adjusted indentation for better readability.